day bai tap ngay 15/03/2022

+++ JPA/thuc_hanh/customer_2/src/main/java/com/example/test/service/ICustomerService.java
@@ -5,7 +5,7 @@ import com.example.test.model.Customer;
 import java.util.List;
 
 public interface ICustomerService {
-    Customer findById(int id);
+    Customer findById(Long id);
     List<Customer> findAll();
     List<Customer>searchByName(String name);
     Customer findOne(Long id);
